moved public static  pages to public paged folder
